Fixes sketchfab embeds
Recent changes to the og:video tag on sketchfab allows us to pass several url parameters instead of just autostart.
For links that whose first parameter is not autostart, the previous code did nothing, causing the embed to autostart.
For links that have several parameters and start with autostart, the `?` is removed from the url, causing a 404.

This change removes only `autostart=1`, removing the autostart behavior and keeping all other behaviors as expected.