FIX - New image URLs. Usernames with `.` characters. Different description formats.
The previous way of obtaining the image (appending `/media/?size=l` to the URL) has been deprecated. Update to use the image URL that is explicitly returned as part of the response payload.

NOTE: this image URL is timestamped and will only work for a limited period of time, so the image may need to be cached locally.

The previous author_name regex failed if the name contained a period character (`.`). Regex updated to allow for those.

Sometimes a description is returned in a different format. Handle those cases.