DEV: Attempt to fetch/parse YouTube ‘embed’ responses
YouTube embed responses appear to give us enough data to construct a Onebox, without requiring EU-based requests to consent to various T&Cs before we can get the result.

Note: This requires a change in the `lazy-yt` plugin in Discourse core for the solution to fully work within the Discourse app.